+++ cassandra/cluster.py
@@ -397,6 +397,9 @@ class Cluster(object):
     """
 
     allow_beta_protocol_version = False
+
+    no_compact = False
+
     """
     Setting true injects a flag in all messages that makes the server accept and use "beta" protocol version.
     Used for testing new protocol features incrementally before the new version is complete.
@@ -771,7 +774,8 @@ class Cluster(object):
                  prepare_on_all_hosts=True,
                  reprepare_on_up=True,
                  execution_profiles=None,
-                 allow_beta_protocol_version=False):
+                 allow_beta_protocol_version=False,
+                 no_compact=False):
         """
         ``executor_threads`` defines the number of threads in a pool for handling asynchronous tasks such as
         extablishing connection pools or refreshing metadata.
@@ -797,6 +801,8 @@ class Cluster(object):
             self.protocol_version = protocol_version
             self._protocol_version_explicit = True
         self.allow_beta_protocol_version = allow_beta_protocol_version
+
+        self.no_compact = no_compact
 
         self.auth_provider = auth_provider
 
@@ -1134,6 +1140,7 @@ class Cluster(object):
         kwargs_dict.setdefault('protocol_version', self.protocol_version)
         kwargs_dict.setdefault('user_type_map', self._user_types)
         kwargs_dict.setdefault('allow_beta_protocol_version', self.allow_beta_protocol_version)
+        kwargs_dict.setdefault('no_compact', self.no_compact)
 
         return kwargs_dict